+++ src/main/java/de/keybird/beagle/JobExecutionManagerConfig.java
@@ -41,6 +41,5 @@ public class JobExecutionManagerConfig {
         final int preferredPoolSize = Math.round(maxHeapSize / 1024f / 1024f / 1024f);
         int poolSize = Math.min(Math.max(minPoolSize, preferredPoolSize), maxPoolSize);
         return poolSize;
-//        return 1; // TODO MVR for now always return 1 as it causes issues :(
     }
 }
