+++ src/main/java/de/keybird/beagle/jobs/JobState.java
@@ -16,11 +16,11 @@
  * along with Beagle. If not, see http://www.gnu.org/licenses/.
  */
 
-package de.keybird.beagle.jobs.xxxx;
+package de.keybird.beagle.jobs;
 
-public enum LogLevel {
-    Info,
-    Success,
-    Warn,
-    Error
+public enum JobState {
+    Pending,
+    Initializing,
+    Running,
+    Completed
 }
