+++ src/main/java/de/keybird/beagle/jobs/execution/ErrorHandler.java
@@ -18,7 +18,7 @@
 
 package de.keybird.beagle.jobs.execution;
 
-import de.keybird.beagle.jobs.xxxx.Job;
+import de.keybird.beagle.jobs.Job;
 
 public interface ErrorHandler<J extends Job> {
     void handle(JobExecutionContext<J> context, Throwable throwable);
