+++ src/main/java/de/keybird/beagle/jobs/execution/IndexJobExecution.java
@@ -40,8 +40,8 @@ import de.keybird.beagle.api.Page;
 import de.keybird.beagle.api.PageState;
 import de.keybird.beagle.elastic.BulkResultWrapper;
 import de.keybird.beagle.elastic.FailedItem;
-import de.keybird.beagle.jobs.xxxx.IndexJob;
-import de.keybird.beagle.jobs.xxxx.LogLevel;
+import de.keybird.beagle.jobs.IndexJob;
+import de.keybird.beagle.jobs.LogLevel;
 import de.keybird.beagle.repository.PageRepository;
 import io.searchbox.client.JestClient;
 import io.searchbox.core.Bulk;
