+++ src/main/java/de/keybird/beagle/jobs/execution/JobExecution.java
@@ -18,7 +18,7 @@
 
 package de.keybird.beagle.jobs.execution;
 
-import de.keybird.beagle.jobs.xxxx.Job;
+import de.keybird.beagle.jobs.Job;
 
 public interface JobExecution<J extends Job> {
     void execute(JobExecutionContext<J> context) throws Exception;
