+++ src/main/java/de/keybird/beagle/jobs/execution/SuccessHandler.java
@@ -18,7 +18,7 @@
 
 package de.keybird.beagle.jobs.execution;
 
-import de.keybird.beagle.jobs.xxxx.Job;
+import de.keybird.beagle.jobs.Job;
 
 public interface SuccessHandler<J extends Job> {
     void handle(JobExecutionContext<J> context);
