+++ src/main/java/de/keybird/beagle/jobs/source/InboxFileSystemSource.java
@@ -16,7 +16,7 @@
  * along with Beagle. If not, see http://www.gnu.org/licenses/.
  */
 
-package de.keybird.beagle.jobs.xxxx;
+package de.keybird.beagle.jobs.source;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -26,6 +26,8 @@ import java.util.Objects;
 import java.util.stream.Collectors;
 
 import de.keybird.beagle.api.DocumentSource;
+import de.keybird.beagle.jobs.Job;
+import de.keybird.beagle.jobs.LogLevel;
 import de.keybird.beagle.jobs.execution.JobExecutionContext;
 
 public class InboxFileSystemSource implements DocumentSource {
