+++ src/main/java/de/keybird/beagle/rest/JobRestController.java
@@ -37,9 +37,9 @@ import de.keybird.beagle.jobs.JobExecutionManager;
 import de.keybird.beagle.jobs.Progress;
 import de.keybird.beagle.jobs.execution.JobExecutionInfo;
 import de.keybird.beagle.jobs.persistence.JobEntity;
-import de.keybird.beagle.jobs.xxxx.DetectJob;
-import de.keybird.beagle.jobs.xxxx.ImportJob;
-import de.keybird.beagle.jobs.xxxx.Job;
+import de.keybird.beagle.jobs.DetectJob;
+import de.keybird.beagle.jobs.ImportJob;
+import de.keybird.beagle.jobs.Job;
 import de.keybird.beagle.repository.DocumentRepository;
 import de.keybird.beagle.repository.JobRepository;
 import de.keybird.beagle.rest.model.JobDTO;
