+++ src/main/java/de/keybird/beagle/rest/model/JobDTO.java
@@ -23,8 +23,8 @@ import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import de.keybird.beagle.jobs.xxxx.JobState;
-import de.keybird.beagle.jobs.xxxx.JobType;
+import de.keybird.beagle.jobs.JobState;
+import de.keybird.beagle.jobs.JobType;
 import de.keybird.beagle.jobs.persistence.JobEntity;
 
 public class JobDTO {
