+++ src/main/java/de/keybird/beagle/services/JobService.java
@@ -37,9 +37,9 @@ import de.keybird.beagle.api.PageState;
 import de.keybird.beagle.jobs.JobExecutionFactory;
 import de.keybird.beagle.jobs.JobExecutionManager;
 import de.keybird.beagle.jobs.persistence.JobEntity;
-import de.keybird.beagle.jobs.xxxx.ImportJob;
-import de.keybird.beagle.jobs.xxxx.IndexJob;
-import de.keybird.beagle.jobs.xxxx.JobType;
+import de.keybird.beagle.jobs.ImportJob;
+import de.keybird.beagle.jobs.IndexJob;
+import de.keybird.beagle.jobs.JobType;
 import de.keybird.beagle.repository.DocumentRepository;
 import de.keybird.beagle.repository.JobRepository;
 import de.keybird.beagle.repository.PageRepository;
