+++ src/test/java/de/keybird/beagle/jobs/JobExecutionManagerTest.java
@@ -47,9 +47,6 @@ import de.keybird.beagle.TestConfig;
 import de.keybird.beagle.WorkingDirectory;
 import de.keybird.beagle.jobs.persistence.ImportJobEntity;
 import de.keybird.beagle.jobs.persistence.JobEntity;
-import de.keybird.beagle.jobs.xxxx.DetectJob;
-import de.keybird.beagle.jobs.xxxx.JobState;
-import de.keybird.beagle.jobs.xxxx.JobType;
 import de.keybird.beagle.repository.DocumentRepository;
 import de.keybird.beagle.repository.JobRepository;
 import de.keybird.beagle.repository.PageRepository;
